Start of Ui Refactoring
