+++ RootofRiches/Ui/DebugWindow.cs
